\name{parse.name.description}
\alias{parse.name.description}
\title{Parse an element containing a mandatory name...}
\usage{parse.name.description(key, rest)}
\description{Parse an element containing a mandatory name
and description (such as \code{@param}).}
\value{A list containing the key, name and
description}
\arguments{\item{key}{the parsing key}
\item{rest}{the expression to be parsed}}
