\name{parser.default}
\alias{parser.default}
\title{Default parser-lookup; if key not found, return...}
\usage{parser.default(table, key, default)}
\description{Default parser-lookup; if key not found, return
the default parser specified.}
\value{The parser}
\arguments{\item{table}{the parser table from which to look}
\item{key}{the key upon which to look}
\item{default}{the parser to return upon unsuccessful
lookup}}
