\name{SH.PP.2.E}
\alias{SH.PP.2.E}
\title{Calculate water vapor partial pressure of an air parcel from specific humidity}
\description{
Calculate water vapor partial pressure of an air parcel given in  (hPa) from the air parcel's specific humidity (kg/kg) and its ambient pressure (hPa).
}
\usage{
SH.PP.2.E(sh, pp=1013.25)
}
\arguments{
  \item{sh}{specific humidity (kg/kg)}
  \item{pp}{ambient pressure (hPa)}
}
\value{
water vapor partial pressure (hPa)
}
\author{Stephan Henne}
\examples{
}
\keyword{manip}
